Added a new class RefactoringMinerWrapper that was extracted from class SmellRefactoredManager.
